vendor/zen: Add custom font package
Fonts taken from dafont.com and free to use for personal use and not
commercial use.

https://www.dafont.com/

According to 18 USCS § 31, term "used for commercial purposes" means the
carriage of persons or property for any fare, fee, rate, charge or other
consideration, or directly or indirectly in connection with any business,
or other undertaking intended for profit.

https://definitions.uslegal.com/u/used-for-commercial-purposes/

Since DU is not and will never be a for profit, we're in the clear :-)

- Fonts xml files generated for these font zips thanks to @USA-RedDragon

Change-Id: I172f0326e1094024ba74a703ab43ace2786e0d46
Signed-off-by: AmolAmrit <amol.amrit03@outlook.com>

CustomFonts: Add more user suggested fonts

-Thanks to CarbonROM team and users

Change-Id: Ic32522883ed35c9798e51fc9de0c0522e3c452b3
Signed-off-by: AmolAmrit <amol.amrit03@outlook.com>

Added new fonts to the custom fonts pack

- ubuntu regular
- ohd5
- firasans
- exo2

all are free to use according to the source i got it from.

Change-Id: I686536678fa0d463890cf444c950b43b3878d1c3

Add Anaheim cuz its sexy

 * Fix name in fonts.xml

Change-Id: I2c16c54de1529cbbd10a01db8f202a04342f1193
Signed-off-by: Shubham Singh <coolsks94@gmail.com>

Vendor:Add some more fonts

 * MontsterratAlternate
 * Barlow
 * Gafata
 * Oswald
 * Raleway
 * Rubik

Change-Id: I6273bd58f49883629ac385a6937957b3d3c47f22
Signed-off-by: Shubham Singh <coolsks94@gmail.com>

Add some of my Favorite fonts

Change-Id: Ic00186b7babce46e14051831604190603da9d621